+++ src/test/java/nl/lexemmens/podman/BuildMojoTest.java
@@ -540,6 +540,61 @@ public class BuildMojoTest extends AbstractMojoTest {
     }
 
     @Test
+    public void testMultiStageBuildWithCustomTagPerStageButStageNameIsNotConfigured() throws MojoExecutionException, IOException, URISyntaxException {
+        URI sampleBuildOutputUri = PushMojoTest.class.getResource("/multistagecontainerfile/samplebuildoutput.txt").toURI();
+        Path sampleBuildOutputPath = Paths.get(sampleBuildOutputUri);
+
+        List<String> buildOutputUnderTest = null;
+        try (Stream<String> buildSampleOutput = Files.lines(sampleBuildOutputPath)) {
+            buildOutputUnderTest = buildSampleOutput.collect(Collectors.toList());
+        }
+
+        Assertions.assertNotNull(buildOutputUnderTest);
+
+        PodmanConfiguration podman = new TestPodmanConfigurationBuilder().setTlsVerify(TlsVerify.FALSE).build();
+        SingleImageConfiguration image = new TestSingleImageConfigurationBuilder("sample")
+                .setContainerfileDir("src/test/resources/multistagecontainerfile")
+                .setTags(new String[]{"0.2.1"})
+                .setCreateLatestTag(false)
+                .setUseCustomImageNameForMultiStageContainerfile(true)
+                .addCustomImageNameForBuildStage("non-existent-phase", "image-name-number-1")
+                .build();
+        configureMojo(podman, image, true, false, false, false, true);
+
+        when(mavenProject.getBuild()).thenReturn(build);
+        when(build.getDirectory()).thenReturn("target");
+        when(serviceHubFactory.createServiceHub(isA(Log.class), isA(MavenProject.class), isA(MavenFileFilter.class), isA(PodmanConfiguration.class), isA(Settings.class), isA(SettingsDecrypter.class))).thenReturn(serviceHub);
+        when(serviceHub.getContainerfileDecorator()).thenReturn(containerfileDecorator);
+        when(serviceHub.getPodmanExecutorService()).thenReturn(podmanExecutorService);
+        when(podmanExecutorService.build(isA(SingleImageConfiguration.class))).thenReturn(buildOutputUnderTest);
+
+        buildMojo.execute();
+
+        // Verify logging
+        verify(log, times(1)).info("Detected multistage Containerfile...");
+
+        // At random verify some lines
+        verify(log, times(1)).debug("Processing line: 'STEP 1: FROM nexus.example:15000/adoptopenjdk/openjdk11:11.0.3 AS base'");
+        verify(log, times(1)).debug("Processing candidate: 'STEP 7: LABEL Build-User=sample-user2 Git-Repository-Url=null'");
+
+        // Verify stage detection
+        verify(log, times(1)).debug("Processing stage in Containerfile: base");
+        verify(log, times(1)).debug("Processing stage in Containerfile: phase");
+        verify(log, times(1)).debug("Processing stage in Containerfile: phase2");
+
+        // Verify hashes for stages
+        verify(log, times(1)).info("Final image for stage base is: 7e72c870614");
+        verify(log, times(1)).info("Final image for stage phase is: 7f55eab001a");
+        verify(log, times(1)).info("Final image for stage phase2 is: d2efc6645cb");
+
+        // Verify warning markers
+        verify(log, times(1)).warn("No image name configured for build stage: phase. Image 7f55eab001a not tagged!");
+        verify(log, times(1)).warn("No image name configured for build stage: phase2. Image d2efc6645cb not tagged!");
+
+        verify(log, times(1)).info("Built container image.");
+    }
+
+    @Test
     public void testMultiStageContainerFileWithMultilineOutputStep() throws MojoExecutionException, IOException, URISyntaxException {
         URI sampleBuildOutputUri = PushMojoTest.class.getResource("/multistagecontainerfile/samplebuildoutput_multiline_step.txt").toURI();
         Path sampleBuildOutputPath = Paths.get(sampleBuildOutputUri);